+++ modules/mex/src/cpp/MxObject.cpp
@@ -26,8 +26,10 @@
 #include <cstring>
 #include "mex.h"
 #include "MxObject.h"
+#include "MxCall.h"
 #include "MxTypes.h"
 #include "ArrayOf.hpp"
+#include "Evaluator.hpp"
 #include "ClassName.hpp"
 #include "MxArrayOf.hpp"
 //=============================================================================
@@ -40,51 +42,40 @@ mxIsClass(const mxArray* pm, const char* classname)
     }
     if (strcmp(classname, "cell") == 0) {
         return (pm->classID == mxCELL_CLASS);
-    }
-    if (strcmp(classname, "char") == 0) {
+    } else if (strcmp(classname, "char") == 0) {
         return (pm->classID == mxCHAR_CLASS);
-    }
-    if (strcmp(classname, "double") == 0) {
+    } else if (strcmp(classname, "double") == 0) {
         return (pm->classID == mxDOUBLE_CLASS);
-    }
-    if (strcmp(classname, "function_handle") == 0) {
+    } else if (strcmp(classname, "function_handle") == 0) {
         return (pm->classID == mxFUNCTION_CLASS);
-    }
-    if (strcmp(classname, "int8") == 0) {
+    } else if (strcmp(classname, "int8") == 0) {
         return (pm->classID == mxINT8_CLASS);
-    }
-    if (strcmp(classname, "int16") == 0) {
+    } else if (strcmp(classname, "int16") == 0) {
         return (pm->classID == mxINT16_CLASS);
-    }
-    if (strcmp(classname, "int32") == 0) {
+    } else if (strcmp(classname, "int32") == 0) {
         return (pm->classID == mxINT32_CLASS);
-    }
-    if (strcmp(classname, "int64") == 0) {
+    } else if (strcmp(classname, "int64") == 0) {
         return (pm->classID == mxINT64_CLASS);
-    }
-    if (strcmp(classname, "logical") == 0) {
+    } else if (strcmp(classname, "logical") == 0) {
         return (pm->classID == mxLOGICAL_CLASS);
-    }
-    if (strcmp(classname, "single") == 0) {
+    } else if (strcmp(classname, "single") == 0) {
         return (pm->classID == mxSINGLE_CLASS);
-    }
-    if (strcmp(classname, "struct") == 0) {
+    } else if (strcmp(classname, "struct") == 0) {
         return (pm->classID == mxSTRUCT_CLASS);
-    }
-    if (strcmp(classname, "uint8") == 0) {
+    } else if (strcmp(classname, "uint8") == 0) {
         return (pm->classID == mxUINT8_CLASS);
-    }
-    if (strcmp(classname, "uint16") == 0) {
+    } else if (strcmp(classname, "uint16") == 0) {
         return (pm->classID == mxUINT16_CLASS);
-    }
-    if (strcmp(classname, "uint32") == 0) {
+    } else if (strcmp(classname, "uint32") == 0) {
         return (pm->classID == mxUINT32_CLASS);
-    }
-    if (strcmp(classname, "uint64") == 0) {
+    } else if (strcmp(classname, "uint64") == 0) {
         return (pm->classID == mxUINT64_CLASS);
-    }
-    if (strcmp(classname, "unknown") == 0) {
+    } else if (strcmp(classname, "unknown") == 0) {
         return (pm->classID == mxUNKNOWN_CLASS);
+    } else {
+        Nelson::ArrayOf nlsArrayOf = Nelson::MxArrayToArrayOf(pm);
+        std::string name = Nelson::ClassName(nlsArrayOf);
+        res = strcmp(classname, name.c_str()) == 0;
     }
     return res;
 }
@@ -146,6 +137,40 @@ mxGetProperty(const mxArray* pa, mwIndex index, const char* propname)
             if (!isSupportedObject) {
                 mexErrMsgTxt(_("mxGetProperty object expected.").c_str());
             }
+            Nelson::Dimensions dims = ptr->getDimensions();
+            if (index > dims.getElementCount()) {
+                return res;
+            }
+            Nelson::ArrayOf obj;
+            if (index == 0) {
+                obj = *ptr;
+            } else {
+                Nelson::ArrayOf* elements = (Nelson::ArrayOf*)ptr->getDataPointer();
+                obj = elements[index];
+            }
+            Nelson::Evaluator* mainEvaluator = (Nelson::Evaluator*)mexGetEvaluator();
+            if (mainEvaluator != nullptr) {
+                Nelson::Context* context = mainEvaluator->getContext();
+                if (context != nullptr) {
+                    Nelson::FunctionDef* funcDef = nullptr;
+                    if (context->lookupFunction("get", funcDef)) {
+                        if (funcDef != nullptr) {
+                            Nelson::ArrayOfVector argIn;
+                            Nelson::ArrayOfVector argOut;
+                            argIn.push_back(obj);
+                            argIn.push_back(Nelson::ArrayOf::characterArrayConstructor(propname));
+                            try {
+                                argOut = funcDef->evaluateFunction(mainEvaluator, argIn, 1);
+                            } catch (Nelson::Exception&) {
+                                return nullptr;
+                            }
+                            if (!argOut.empty()) {
+                                res = Nelson::ArrayOfToMxArray(argOut[0], pa->interleavedcomplex);
+                            }
+                        }
+                    }
+                }
+            }
         }
     }
     return res;
@@ -162,6 +187,36 @@ mxSetProperty(mxArray* pa, mwIndex index, const char* propname, const mxArray* v
                     = ptr->isClassStruct() || ptr->isHandle() || ptr->isGraphicObject();
                 if (!isSupportedObject) {
                     mexErrMsgTxt(_("mxSetProperty object expected.").c_str());
+                }
+                Nelson::Dimensions dims = ptr->getDimensions();
+                if (index <= dims.getElementCount()) {
+                    Nelson::ArrayOf obj;
+                    if (index == 0) {
+                        obj = *ptr;
+                    } else {
+                        Nelson::ArrayOf* elements = (Nelson::ArrayOf*)ptr->getDataPointer();
+                        obj = elements[index];
+                    }
+                    Nelson::Evaluator* mainEvaluator = (Nelson::Evaluator*)mexGetEvaluator();
+                    if (mainEvaluator != nullptr) {
+                        Nelson::Context* context = mainEvaluator->getContext();
+                        if (context != nullptr) {
+                            Nelson::FunctionDef* funcDef = nullptr;
+                            if (context->lookupFunction("set", funcDef)) {
+                                if (funcDef != nullptr) {
+                                    Nelson::ArrayOfVector argIn;
+                                    Nelson::ArrayOfVector argOut;
+                                    argIn.push_back(obj);
+                                    argIn.push_back(
+                                        Nelson::ArrayOf::characterArrayConstructor(propname));
+                                    argIn.push_back(Nelson::MxArrayToArrayOf(value));
+                                    try {
+                                        argOut = funcDef->evaluateFunction(mainEvaluator, argIn, 0);
+                                    } catch (Nelson::Exception&) { }
+                                }
+                            }
+                        }
+                    }
                 }
             }
         }